Move types into own package to import in api_tests
